removed ability for non-stripped log to be tapped
