+++ src/main/java/io/github/darealturtywurty/turtychemistry/common/block/treeblocks/rubbertree/RubberTreeBlock.java
@@ -4,8 +4,6 @@ import io.github.darealturtywurty.turtychemistry.core.init.BlockInit;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.properties.BooleanProperty;
-import net.minecraft.world.level.block.state.properties.IntegerProperty;
 import net.minecraftforge.common.ToolAction;
 import net.minecraftforge.common.ToolActions;
 import org.jetbrains.annotations.NotNull;
@@ -14,8 +12,6 @@ import org.jetbrains.annotations.Nullable;
 import java.util.concurrent.ThreadLocalRandom;
 
 public final class RubberTreeBlock extends RubberTreeBaseBlock {
-    public static final IntegerProperty RUBBER_IN_TREE = IntegerProperty.create("rubber_in_tree", 0, 5);
-    public static final BooleanProperty HAS_RUBBER = BooleanProperty.create("has_rubber");
 
     public RubberTreeBlock(final Properties blockProperty) {
         super(blockProperty);
@@ -29,7 +25,6 @@ public final class RubberTreeBlock extends RubberTreeBaseBlock {
             return BlockInit.RUBBER_TREE_BLOCK_STRIPPED.get().defaultBlockState().setValue(RUBBER_IN_TREE, state.getValue(RUBBER_IN_TREE))
                     .setValue(HAS_RUBBER, state.getValue(HAS_RUBBER)).setValue(AXIS, state.getValue(AXIS));
         }
-
         return super.getToolModifiedState(state, context, toolAction, simulate);
     }
